% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nolock.R
\docType{data}
\name{stringr}
\alias{stringr}
\title{*---------------------------------------------------------------------------*
*         Kod dodający WITH (NOLOCK) po tabelach w zapytaniach SQL          *
*---------------------------------------------------------------------------*}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
stringr
}
\description{
*---------------------------------------------------------------------------*
*         Kod dodający WITH (NOLOCK) po tabelach w zapytaniach SQL          *
*---------------------------------------------------------------------------*
}
\keyword{datasets}
